removed support for legacy tests
