--- json/integrationTests/example.test.json
@@ -1,109 +1,0 @@
-{
-    "ERC20_IELE": {
-        "pre": {
-            "0xa94f5374fce5edbc8e2a8697c15331677e6ebf0b000000000000000000000000": {
-                "nonce": "0x00",
-                "balance": "0xe8d4a51000"
-            },
-            "0x1000000000000000000000000000000000000000000000000000000000000000": {
-                "nonce": "0x00",
-                "balance": "0x00",
-                "storage": {
-                    "0x19efaebcc296cffac396adb4a60d54c05eff43926a6072498a618e943908efe1": "0x2710",
-                    "0x0000000000000000000000000000000000000000000000000000000000000000": "0x2710"
-                },
-                "code": "file:smart-contract.wasm"
-            }
-        },
-        "blocks": [
-            {
-                "results": [
-                    {
-                        "out": [],
-                        "status": "",
-                        "logs": [
-                            {
-                                "address": "0x1000000000000000000000000000000000000000000000000000000000000000",
-                                "endpoint": "str:transferFrom",
-                                "topics": [
-                                    "0xf099cd8bde557814842a3121e8ddfd433a539b8c9f14bf31ebf108d12e6196e9",
-                                    "0xa94f5374fce5edbc8e2a8697c15331677e6ebf0b000000000000000000000000",
-                                    "0x1234123400000000000000000000000000000000000000000000000000000004"
-                                ],
-                                "data": []
-                            }
-                        ],
-                        "gas": "0x0f7b9d",
-                        "refund": ""
-                    },
-                    {
-                        "out": [],
-                        "status": "",
-                        "logs": [],
-                        "gas": "*",
-                        "refund": ""
-                    }
-                ],
-                "transactions": [
-                    {
-                        "nonce": "0x00",
-                        "function": "transferToken",
-                        "gasLimit": "0x100000",
-                        "value": "0x00",
-                        "to": "0x1000000000000000000000000000000000000000000000000000000000000000",
-                        "arguments": [
-                            "0x1234123400000000000000000000000000000000000000000000000000000004",
-                            "0x00"
-                        ],
-                        "gasPrice": "0x01",
-                        "from": "0xa94f5374fce5edbc8e2a8697c15331677e6ebf0b000000000000000000000000"
-                    },
-                    {
-                        "nonce": "0x00",
-                        "function": "",
-                        "gasLimit": "0x100000",
-                        "value": "0x00",
-                        "to": "",
-                        "arguments": [
-                            "0x1234123400000000000000000000000000000000000000000000000000000004",
-                            "0x00"
-                        ],
-                        "contractCode": "``contract code to deploy",
-                        "gasPrice": "0x01",
-                        "from": "0xa94f5374fce5edbc8e2a8697c15331677e6ebf0b000000000000000000000000"
-                    }
-                ],
-                "blockHeader": {
-                    "gasLimit": "0x174876e800",
-                    "number": "0x01",
-                    "difficulty": "0x020000",
-                    "timestamp": "0x03e8",
-                    "coinbase": "0x2adc25665018aa1fe0e6bc666dac8fc2697ff9ba"
-                }
-            }
-        ],
-        "network": "VM",
-        "blockHashes": [
-            "0x24a30e4305ac41674b26493c800c05f507e98d3b8bceb0a314f9b9bc43622736",
-            "0x00"
-        ],
-        "postState": {
-            "0xa94f5374fce5edbc8e2a8697c15331677e6ebf0b000000000000000000000000": {
-                "nonce": "0x01",
-                "balance": "0xe8d4951000",
-                "storage": {},
-                "code": ""
-            },
-            "0x1000000000000000000000000000000000000000000000000000000000000000": {
-                "nonce": "0x00",
-                "balance": "0x00",
-                "storage": {
-                    "0x19efaebcc296cffac396adb4a60d54c05eff43926a6072498a618e943908efe1": "0x2710",
-                    "0x0000000000000000000000000000000000000000000000000000000000000000": "0x2710"
-                },
-                "code": "file:smart-contract.wasm",
-                "asyncCallData": "``someAsync@call@data"
-            }
-        }
-    }
-}
